Ajout de nouveau tests fonctionnels

+++ src/main/java/com/cesi/ressourcesrelationnelles/service/RessourceService.java
@@ -1,9 +1,12 @@
 package com.cesi.ressourcesrelationnelles.service;
 
 import com.cesi.ressourcesrelationnelles.domain.Resource;
+import com.cesi.ressourcesrelationnelles.exception.ResourceAlreadyExistException;
 import com.cesi.ressourcesrelationnelles.exception.ResourceNotFoundException;
+import com.cesi.ressourcesrelationnelles.exception.ResourceNotValidException;
 import com.cesi.ressourcesrelationnelles.repository.RessourceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -25,7 +28,6 @@ public class RessourceService {
         for (Resource item : iterable) {
             resources.add(item);
         }
-        //List<Ressource> resources = resourceRepository.findAll();
         return resources;
     }
 
@@ -52,19 +54,47 @@ public class RessourceService {
         throw new ResourceNotFoundException();
     }
 
-    public Resource createResource(Resource resource){
-        return resourceRepository.save(resource);
+    public Resource createResource(Resource resource) throws ResourceNotValidException, ResourceAlreadyExistException, ResourceNotFoundException {
+        if(resource.getId() == null || resource.getTitle() == null || resource.getPublishDate() == null || resource.getUrlContent() == null || resource.getViewNumber() == null || resource.getAuthor() == null || resource.getType() == null || resource.getCategory() == null){
+            throw new ResourceNotValidException();
+        }
+        if(this.idExists(resource.getId())){
+            throw new ResourceAlreadyExistException();
+        }
+        try {
+            return resourceRepository.save(resource);
+        }catch(JpaObjectRetrievalFailureException e){
+            throw new ResourceNotValidException();
+        }
     }
 
-    public Resource updateResource(Resource resource) throws ResourceNotFoundException {
-        if(resourceRepository.findById(resource.getId()).isPresent()){
+    private boolean idExists(Long id) {
+        Optional<Resource> res = resourceRepository.findById(id);
+        if(res.isPresent()){
+            return true;
+        }
+        return false;
+    }
+
+    public Resource updateResource(Resource resource) throws ResourceNotFoundException, ResourceNotValidException {
+        Optional<Resource> resOpt = resourceRepository.findById(resource.getId());
+        if(resOpt.isPresent()){
+            Resource old = resOpt.get();
+            if(old.getPublishDate().compareTo(resource.getPublishDate()) != 0){
+                throw new ResourceNotValidException();
+            }
             return resourceRepository.save(resource);
         }
         throw new ResourceNotFoundException();
     }
 
-    public void deleteResource(long id){
-        resourceRepository.deleteById(id);
+    public void deleteResource(long id) throws ResourceNotFoundException {
+        if(this.idExists(id)) {
+            resourceRepository.deleteById(id);
+        }
+        else {
+            throw new ResourceNotFoundException();
+        }
     }
 
     public List<Resource> listOrderByDateDesc() {
